adjustments to DeployDialog according to react-refresh lint rule
