+++ frontend/app/src/components/StreamlitDialog/DeployDialog/DeployDialog.tsx
@@ -20,10 +20,8 @@ import { StyledAction, StyledBody } from "baseui/card"
 
 import { BaseButton, BaseButtonKind, GitInfo, IGitInfo } from "@streamlit/lib"
 import { MetricsManager } from "@streamlit/app/src/MetricsManager"
-import {
-  DialogType,
-  PlainEventHandler,
-} from "@streamlit/app/src/components/StreamlitDialog/StreamlitDialog"
+import { PlainEventHandler } from "@streamlit/app/src/components/StreamlitDialog/StreamlitDialog"
+import { DialogType } from "@streamlit/app/src/components/StreamlitDialog/constants"
 import { AppContext } from "@streamlit/app/src/components/AppContext"
 import StreamlitLogo from "@streamlit/app/src/assets/svg/logo.svg"
 import Rocket from "@streamlit/app/src/assets/svg/rocket.svg"
