+++ frontend/app/src/components/StreamlitDialog/index.tsx
@@ -13,5 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-export * from "./StreamlitDialog"
+export { StreamlitDialog } from "./StreamlitDialog"
+export type { DialogProps, PlainEventHandler } from "./StreamlitDialog"
